+++ packages/lazy-table-renderer/vite.config.js
@@ -3,13 +3,13 @@ import { resolve } from 'path'
 import dts from 'vite-plugin-dts'
 import vue from '@vitejs/plugin-vue'
 
-export default defineConfig(() => {
+export default defineConfig(({ command, mode, ssrBuild }) => {
   return {
     plugins: [
       vue(),
       dts({
-        include: ['src'],
-        outDir: 'dist/types',
+        include: ['src/**/*.{js,ts,vue,d.ts}'],
+        outDir: 'dist',
         staticImport: false,
         insertTypesEntry: true,
         rollupTypes: true
@@ -18,10 +18,9 @@ export default defineConfig(() => {
     publicDir: false,
     build: {
       lib: {
-        entry: resolve(__dirname, 'src/index.ts'),
+        entry: resolve(__dirname, 'src/index.js'),
         name: 'LazyTableRenderer',
-        fileName: (format) => `lazy-table-renderer.${format}.js`,
-        formats: ['es', 'umd']
+        fileName: (format) => `lazy-table-renderer.${format}.js`
       },
       rollupOptions: {
         external: ['vue'],