+++ src/components/pivottable-ui/VPivottableUi.vue
@@ -131,12 +131,13 @@
   </table>
 </template>
 
-<script setup lang="ts">
-import { aggregators, PivotData, sortAs } from '@/helper'
+<script setup>
+import { defaultProps, PivotData, sortAs } from '@/helper'
 import VRendererCell from './VRendererCell.vue'
 import VAggregatorCell from './VAggregatorCell.vue'
 import VDragAndDropCell from './VDragAndDropCell.vue'
 import VPivottable from '../pivottable/VPivottable.vue'
+import TableRenderer from '../pivottable/renderer/index'
 import { computed, watch } from 'vue'
 import {
   usePropsState,
@@ -144,36 +145,38 @@ import {
   usePivotUiState,
   provideFilterBox
 } from '@/composables'
-import { DefaultPropsType } from '@/types'
-import { locales } from '@/helper'
 
-const props = withDefaults(
-  defineProps<
-    DefaultPropsType & {
-      hiddenAttributes?: string[]
-      hiddenFromAggregators?: string[]
-      hiddenFromDragDrop?: string[]
-      restrictedFromDragDrop?: string[]
-      menuLimit?: number
-      pivotModel?: any
-      hideFilterBoxOfUnusedAttributes?: boolean
-    }
-  >(),
-  {
-    aggregators: () => aggregators,
-    hiddenAttributes: () => [],
-    hiddenFromAggregators: () => [],
-    hiddenFromDragDrop: () => [],
-    restrictedFromDragDrop: () => [],
-    menuLimit: 500,
-    hideFilterBoxOfUnusedAttributes: false,
-    rowOrder: 'key_a_to_z',
-    colOrder: 'key_a_to_z',
-    languagePack: () => locales,
-    locale: 'en'
+const props = defineProps({
+  ...defaultProps,
+  hiddenAttributes: {
+    type: Array,
+    default: () => []
+  },
+  hiddenFromAggregators: {
+    type: Array,
+    default: () => []
+  },
+  hiddenFromDragDrop: {
+    type: Array,
+    default: () => []
+  },
+  restrictedFromDragDrop: {
+    type: Array,
+    default: () => []
+  },
+  menuLimit: {
+    type: Number,
+    default: 500
+  },
+  pivotModel: {
+    type: Object,
+    default: () => ({})
+  },
+  hideFilterBoxOfUnusedAttributes: {
+    type: Boolean,
+    default: false
   }
-)
-
+})
 const {
   state,
   localeStrings,
@@ -202,7 +205,7 @@ const { allFilters, materializedInput } = useMaterializeInput(
 )
 
 const rendererItems = computed(() =>
-  Object.keys(state.renderers).length ? state.renderers : {}
+  Object.keys(state.renderers).length ? state.renderers : TableRenderer
 )
 const aggregatorItems = computed(() => state.aggregators)
 const rowAttrs = computed(() => {
@@ -235,7 +238,7 @@ const unusedAttrs = computed(() => {
         !state.hiddenAttributes.includes(e) &&
         !state.hiddenFromDragDrop.includes(e)
     )
-    .sort(sortAs(pivotUiState.unusedOrder))
+    .sort(sortAs(state.unusedOrder))
 })
 
 const pivotData = computed(() => new PivotData(state))
@@ -261,13 +264,12 @@ const pivotProps = computed(() => ({
   rowOrder: state.rowOrder,
   colOrder: state.colOrder,
   tableMaxWidth: state.tableMaxWidth,
-  localeStrings: localeStrings.value,
-  menuLimit: props.menuLimit
+  localeStrings: localeStrings.value
 }))
 
-onUpdateUnusedOrder(unusedAttrs.value)
+onUpdateUnusedOrder(props.unusedAttrs)
 
-provideFilterBox(pivotProps.value)
+provideFilterBox(props)
 
 watch(
   [allFilters, materializedInput],