+++ src/components/pivottable/renderer/TSVExportRenderers.vue
@@ -6,13 +6,22 @@
   />
 </template>
 
-<script setup lang="ts">
+<script setup>
+import { defaultProps, PivotData } from '@/helper'
 import { computed } from 'vue'
-import { usePivotData } from '@/composables'
-import { DefaultPropsType } from '@/types'
 
-const props = defineProps<DefaultPropsType>()
-const { pivotData, error } = usePivotData(props)
+const props = defineProps({
+  ...defaultProps
+})
+
+const pivotData = computed(() => {
+  try {
+    return new PivotData(props)
+  } catch (err) {
+    console.error(err.stack)
+    return null
+  }
+})
 
 const rowKeys = computed(() => pivotData.value?.getRowKeys() || [])
 const colKeys = computed(() => pivotData.value?.getColKeys() || [])
@@ -23,25 +32,22 @@ const styles = {
 }
 
 const headerRow = computed(() => {
-  if (error.value || !pivotData.value) return []
   const header = [...pivotData.value.props.rows]
 
   if (colKeys.value.length === 1 && colKeys.value[0].length === 0) {
     header.push(props.aggregatorName)
   } else {
-    colKeys.value.forEach((c: any[]) => header.push(c.join('-')))
+    colKeys.value.forEach((c) => header.push(c.join('-')))
   }
 
   return header
 })
 
 const result = computed(() => {
-  if (error.value || !pivotData.value) return []
-
-  const data = rowKeys.value.reduce((acc: any[], r: any[]) => {
+  const data = rowKeys.value.reduce((acc, r) => {
     const row = [...r]
-    colKeys.value.forEach((c: any[]) =>
-      row.push(pivotData.value?.getAggregator(r, c).value() || '')
+    colKeys.value.forEach((c) =>
+      row.push(pivotData.value.getAggregator(r, c).value() || '')
     )
     return [...acc, row]
   }, [])