+++ src/helper/index.js
@@ -1,0 +1,3 @@
+export { default as defaultProps } from './defaultProps.js'
+export { redColorScaleGenerator } from './redColorScaleGenerator.js'
+export * from './utilities.js'