+++ src/helper/redColorScaleGenerator.js
@@ -1,0 +1,8 @@
+export function redColorScaleGenerator(values) {
+  const min = Math.min.apply(Math, values)
+  const max = Math.max.apply(Math, values)
+  return (x) => {
+    const nonRed = 255 - Math.round((255 * (x - min)) / (max - min))
+    return { backgroundColor: `rgb(255,${nonRed},${nonRed})` }
+  }
+}